Force LC_NUMERIC=C in the environment
This is a temporary workaround to address a change in apt 1.2.15.

Progress from dpkg/apt used to be in the form of 0.0000, no matter what locale was used.
Since version 1.2.15, it now follows LC_NUMERIC. So in French, Swedish, Russian etc.. it's now in the form of 0,0000.

This causes an exception in python-apt (version 1.1.0~beta1build1) and causes cache commits to fail.

In python-apt's base.py, calling "float(percent)" will fail if it contains a comma.