_flatpak.py: When fetch_remote_ref_sync() fails, fallback to building a synthetic RemoteRef for the runtime.  Remove version comparison from the theme filter.
The synthetic ref creation was already in place for RelatedRefs, as
fetch_remote_ref_sync would frequently fail.  As mentioned in FIXME
comments, some simplification can probably be done where we use the
list_remote_related_refs_async method - there's no need to make another
set of requests for the RemoteRef objects of these related refs.
Our primary purpose here is to confirm they exist.  We can safely
assume they do if they're returned by fetch_remote_ref_sync() and
use the refs in that returned list and jump directly to constructing
a synthetic remote ref (to preserve compatibility with later functions)
or modifying later functions to handle a RelatedRef to get download/
install size info.

The theme/gtk change addresses the fact that freedesktop and Gtk
runtime versions are not the same or related, so it's not a good idea
to compare the theme versions (which match gtk versions) with runtime
versions.  Since there are currently only gtk 3.22 themes available,
it's probably safe to just grab the highest version for now.