_apt.py: don't filter out package 'librecad'.
When constructing our package cache, we try to filter out many
types of packages that a user ordinarily wouldn't need, and only
serve to bog everything down.  Packages that start with lib fit
that description, and there are a lot of them.  Unfortunately,
though we exclude libreoffice packages here, we missed librecad.

Fixes #204