mintinstall.py: search for a list of terms anywhere in a package name, rather than trying to exactly match the search string.
By searching for a literal match in a package's name, we would be
stymied by naming practices that differ between flatpak and apt, along
with the fact that a user might search for a human-readable name, not
one containing dashes or period separators.

Instead, split the search string by whitespace characters into an array,
and look for these substrings within a package's name.