mintinstall.py: Don't try to access pkginfo before it's been set.
This can happen on the first iteration of the loop only. Subsequent
iterations end up using the previous iteration's value of pkginfo,
which is wrong, but allows the loop to avoid an UnboundLocalError
and gets caught as a ValueError instead.

We don't need that code anyhow, as the removal action is performed
later in the loop.

Fixes #314