Add setting to filter apt/flatpak results for the same application
When enabled, search results for applications that are provided as both
apt and flatpak packages will be filtered to only show the user's
preferred package type. As before, the type can still be changed on the
package details page before installation. Applications that only have a
single package will still be shown regardless of type.