Add a setting to filter apt/Flatpak results for the same application, with a fix for search sorting (#398)
* Add setting to filter apt/flatpak results for the same application

When enabled, search results for applications that are provided as both
apt and flatpak packages will be filtered to only show the user's
preferred package type. As before, the type can still be changed on the
package details page before installation. Applications that only have a
single package will still be shown regardless of type.

* Fix search sorting when display name matches but Flatpak ID does not

Previously, if a Flatpak package's display name and summary or
description matched the search terms, but the ID did not, then it would
have been ranked as if there was a match in the summary or description
but not in the name.

* Move all search preferences to a submenu