Use https for debian.net screenshot lookups, improve screenshots for deb packages.
Debian screenshots relative paths now start with /shrine/screenshot
instead of /screenshots.

Adding this results in many more debian package screenshots (We were only
ever getting a single one per package, if that). I'm not sure if
the old path is still used anywhere, so it remains just in case.